crowd-server/requirements/production.txt
-r base.txt
amqp==2.6.1
asgiref==3.5.2
attrs==21.4.0
billiard==3.6.4.0
celery==4.4.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
django-celery-beat==2.1.0
django-timezone-field==4.2.3
flower==0.9.7
humanize==4.1.0
iniconfig==1.1.1
kombu==4.6.11
packaging==21.3
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.29
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
tzdata==2022.1
vine==1.3.0
wcwidth==0.2.5
